Feat: sidebar, KPIs and tabs
